refactor: 패키지 구조 변경 및 기물 클래스 분리
