--- src/main/java/softeer2nd/Pawn.java
@@ -1,15 +1,0 @@
-package softeer2nd;
-
-public class Pawn {
-    Color color;
-
-    Pawn(Color color) {
-        this.color = color;
-    }
-    Pawn() {
-        this.color = Color.WHITE;
-    }
-    Color getColor() {
-        return color;
-    }
-}
