+++ src/main/java/softeer2nd/chess/Main.java
@@ -1,4 +1,4 @@
-package softeer2nd;
+package softeer2nd.chess;
 
 public class Main {
     public static void main(String[] args) {
